refactor: remove docking system, simplify keybinds, and fix process lifecycle
This change is a cleanup and simplification pass over Halley’s input,
configuration model, and window-management behavior.

Major changes:

- Remove docking entirely
  - Deleted docking core module and related rendering (preview, dock-pair)
  - Removed all dock-specific state, overlap handling, and decay logic
  - Simplified focus, carry, and surface lifecycle paths by eliminating
    dock relationship preservation
  - Surfaces are now treated purely as independent nodes again

- Simplify keybinding system
  - Removed legacy/dev keybind fields from RuntimeTuning and Keybinds
  - Deleted legacy keybind parser and fallback paths
  - Input dispatch now relies solely on explicit binding tables
  - Updated config terminology (`docking.*` → `field.gap`)
  - Example config updated to reflect explicit binding model

- Fix spawned process lifecycle
  - Track spawned children in `spawned_children`
  - Include autostart commands in tracked children
  - Explicitly reap children via `try_wait` in backend loops
  - Improves PID tracking and ensures proper cleanup on shutdown

Other changes:

- Logging now reports configured keybind counts instead of legacy fields
- Removed dev-only keybind tuning/movement behavior
- Cleaned up signal handler registration in `run/mod.rs`

Net effect: reduces complexity by removing docking and legacy config paths,
standardizing on explicit bindings, and improving process management.
