--- crates/halley-config/src/legacy.rs
@@ -1,132 +1,0 @@
-use std::collections::HashMap;
-
-pub fn parse_legacy_keybinds(content: &str) -> HashMap<String, String> {
-    let mut out = HashMap::new();
-    let mut in_block = false;
-    let mut depth = 0usize;
-
-    for raw in content.lines() {
-        let trimmed = raw.trim();
-        if trimmed.is_empty() {
-            continue;
-        }
-        if !in_block {
-            if trimmed.eq_ignore_ascii_case("keybinds:") {
-                in_block = true;
-                depth = 1;
-            }
-            continue;
-        }
-
-        if trimmed.eq_ignore_ascii_case("end") {
-            depth = depth.saturating_sub(1);
-            if depth == 0 {
-                in_block = false;
-            }
-            continue;
-        }
-
-        if trimmed.ends_with(':') {
-            depth = depth.saturating_add(1);
-            continue;
-        }
-
-        if depth != 1 {
-            continue;
-        }
-
-        if let Some((k, v)) = parse_legacy_keybind_line(trimmed) {
-            out.insert(k, v);
-        }
-    }
-
-    out
-}
-
-fn parse_legacy_keybind_line(line: &str) -> Option<(String, String)> {
-    let mut clean = String::with_capacity(line.len());
-    let mut in_quotes = false;
-    for ch in line.chars() {
-        if ch == '"' {
-            in_quotes = !in_quotes;
-            clean.push(ch);
-            continue;
-        }
-        if ch == '#' && !in_quotes {
-            break;
-        }
-        clean.push(ch);
-    }
-    let tokens = parse_quoted_tokens(clean.trim());
-    if tokens.len() != 2 {
-        return None;
-    }
-    Some((tokens[0].clone(), tokens[1].clone()))
-}
-
-fn parse_quoted_tokens(input: &str) -> Vec<String> {
-    let mut out = Vec::new();
-    let mut cur = String::new();
-    let mut in_quotes = false;
-    let mut escaped = false;
-
-    for ch in input.chars() {
-        if escaped {
-            cur.push(ch);
-            escaped = false;
-            continue;
-        }
-        if in_quotes && ch == '\\' {
-            escaped = true;
-            continue;
-        }
-        if ch == '"' {
-            in_quotes = !in_quotes;
-            continue;
-        }
-        if !in_quotes && ch.is_ascii_whitespace() {
-            if !cur.is_empty() {
-                out.push(std::mem::take(&mut cur));
-            }
-            continue;
-        }
-        cur.push(ch);
-    }
-    if !cur.is_empty() {
-        out.push(cur);
-    }
-    out
-}
-
-pub fn strip_legacy_keybind_block(content: &str) -> String {
-    let mut out = String::with_capacity(content.len());
-    let mut in_block = false;
-    let mut depth = 0usize;
-
-    for raw in content.lines() {
-        let trimmed = raw.trim();
-        if !in_block {
-            if trimmed.eq_ignore_ascii_case("keybinds:") {
-                in_block = true;
-                depth = 1;
-                continue;
-            }
-            out.push_str(raw);
-            out.push('\n');
-            continue;
-        }
-
-        if trimmed.eq_ignore_ascii_case("end") {
-            depth = depth.saturating_sub(1);
-            if depth == 0 {
-                in_block = false;
-            }
-            continue;
-        }
-        if trimmed.ends_with(':') {
-            depth = depth.saturating_add(1);
-        }
-    }
-
-    out
-}
