+++ crates/halley-config/src/lib.rs
@@ -1,6 +1,5 @@
 pub mod keybinds;
 pub mod layout;
-pub mod legacy;
 pub mod parse;
 
 pub use keybinds::{
