+++ crates/halley-core/src/lib.rs
@@ -2,7 +2,6 @@ pub mod bearings;
 pub mod cluster;
 pub mod cluster_policy;
 pub mod decay;
-pub mod docking;
 pub mod field;
 pub mod focus;
 pub mod tiling;
