+++ crates/halley-wl/src/render/mod.rs
@@ -5,7 +5,6 @@ use halley_core::viewport::{FocusRing, FocusZone, Viewport};
 mod anim_utils;
 mod cursor_render;
 mod cursor_theme;
-mod dock_render;
 mod frame_render;
 mod layer_render;
 mod node_render;
