+++ crates/halley-wl/src/state/mod.rs
@@ -563,9 +563,6 @@ impl HalleyWlState {
             },
             &mut self.cluster_form_state,
         );
-        if !self.suspend_state_checks && self.resize_active.is_none() {
-            self.enforce_docked_pairs();
-        }
         self.enforce_single_primary_active_unit(focus_ring);
         if !self.suspend_state_checks && self.resize_active.is_none() {
             self.resolve_surface_overlap();
